+++ opencog/CMakeLists.txt
@@ -15,7 +15,6 @@ IF (HAVE_ATOMSPACE)
 	ADD_SUBDIRECTORY (eval)
 	ADD_SUBDIRECTORY (query)
 	ADD_SUBDIRECTORY (persist)
-	ADD_SUBDIRECTORY (executioncontext)
 ENDIF (HAVE_ATOMSPACE)
 
 # Extension language support
